implemented the "Finding All the Corners" agent and passed all tests
The correct representation of the probelm (state) is critical!! First, I used the path as the state (line 338) and goal test is to see if the path contains all corners, but it can never end the search! After seeking help in slack, I changed the state representation to (state, visited_corner_list) and it works right away!